backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
sqlalchemy
psycopg2-binary
slowapi
